Separated level from game class. Can now create new class for each level. Level01 and Level02 are test levels. Added folder to data for all the levels.
